Refactor of functional tests: Decouple budget functions in a new file
